% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_reg.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Ridge regression results with a manually selected k value}
\usage{
ridge_reg(x, y, k)
}
\arguments{
\item{x}{Explanatory variables (Dataframe, matrix)}

\item{y}{Dependent variables (Dataframe, vector)}

\item{k}{Ridge parameter}
}
\value{
A list of lists
}
\description{
Ridge regression with a manually selected k value
}
\examples{
library("mctest")
x <- Hald[,-1]
y <- Hald[,1]
k <- 0.1
ridge_reg(x,y,k)

library(isdals)
data(bodyfat)
x <- bodyfat[,-1]
y <- bodyfat[,1]
k <- 0.1
ridge_reg(x,y,k)
}
