% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_k.R
\name{ridgereg_k}
\alias{ridgereg_k}
\title{Ridge regression using alternative techniques}
\usage{
ridgereg_k(x, y, a, b)
}
\arguments{
\item{x}{A Dataframe. Input variables for regression model. It should minium contain two columns.}

\item{y}{Output variable for regression model}

\item{a}{Lower boundary for bias}

\item{b}{Upper boundary for bias}
}
\value{
A list of lists
}
\description{
Ridge regression using alternative techniques.
}
\examples{
library(isdals)
data(bodyfat)
x=bodyfat[,-1]
y=bodyfat[,1]
}
