% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_k.R
\name{ridgereg_k}
\alias{ridgereg_k}
\title{Ridge regression results with an automatically selected k value}
\usage{
ridgereg_k(x, y, a, b)
}
\arguments{
\item{x}{Explanatory variables (Dataframe, matrix)}

\item{y}{Dependent variables (Dataframe, vector)}

\item{a}{Lower bound of k}

\item{b}{Upper bound of k}
}
\value{
A list of lists
}
\description{
Ridge regression with a selected k value
}
\examples{
library("mctest")
x <- Hald[,-1]
y <- Hald[,1]
ridgereg_k(x,y,a=0,b=1)

library(isdals)
data(bodyfat)
x <- bodyfat[,-1]
y <- bodyfat[,1]
ridgereg_k(x,y,a=0,b=1)
}
